Move ReportTable component to separate file.
